% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_funcs.R
\name{error_funcs}
\alias{error_funcs}
\alias{dev_poiss}
\alias{dev_gamma}
\alias{wgt_mse}
\alias{mcc}
\title{Error functions}
\usage{
error_funcs(ytrue, ypred, wcase = rep(1, length(ytrue)), method)

dev_poiss(ytrue, ypred, wcase)

dev_gamma(ytrue, ypred, wcase)

wgt_mse(ytrue, ypred, wcase)

mcc(ytrue, ypred, wcase)
}
\arguments{
\item{ytrue}{numeric vector of the true target values.}

\item{ypred}{numeric vector of the predicted target values.}

\item{wcase}{numeric vector of case weights.}

\item{method}{string specifying the type of random forest. One of the
following: 'anova', 'class', 'poisson', 'gamma' or 'lognormal'.}
}
\description{
The group of loss functions used to calculate OOB errors.
}
\section{Functions}{
\itemize{
\item \code{dev_poiss}: Poisson deviance loss function for Poisson regression

\item \code{dev_gamma}: Gamma deviance loss function for gamma regression

\item \code{wgt_mse}: Weighted mean squared error loss function for anova and lognormal regression

\item \code{mcc}: Matthews correlation coefficient for binary classification
}}

\keyword{internal}
