% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_rforest.R
\name{importance_rforest}
\alias{importance_rforest}
\title{Variable importance in a random forest}
\usage{
importance_rforest(object)
}
\arguments{
\item{object}{fitted model object from the class \code{rforest}.}
}
\value{
data frame with one row for each variable and four columns:
  \describe{\item{variable}{the name of the variable.} \item{importance}{the
  average importance score over all the individual trees.}
  \item{scale_sum}{scaled scores which sum to one.} \item{scale_max}{scaled
  scores such that the maximum value is equal to one.}}
}
\description{
This function acts as a user-friendly interface for the variable importance
scores in a random forest based on individual \code{\link{rpart}} trees.
}
