% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rforest.R
\name{predict.rforest}
\alias{predict.rforest}
\title{Predict a random forest}
\usage{
\method{predict}{rforest}(object, newdata)
}
\arguments{
\item{object}{fitted model object from the class \code{rforest}.}

\item{newdata}{data frame containing the observations to predict. This
argument can only be missing when the random forest in \code{object} is
trained with \code{keep_data = TRUE}. In that case, the original training
data will be used to generate predictions.}
}
\value{
numeric vector with the averaged predictions (for regression) or the
  majority vote (for classification) of the individual trees.
}
\description{
This function obtains predictions from a random forest based on aggregating
the predictions from individual \code{\link{rpart}} trees. A majority vote is
taken for binary classification trees, while the predictions are averaged for
normal, poisson, gamma and lognormal regression trees.
}
